+++ R/datahash.R
@@ -8,29 +8,39 @@
 #' @family internal
 #' @importFrom assertthat assert_that
 #' @importFrom git2r hash
+#' @importFrom utils file_test
 datahash <- function(file) {
-  chunk_size <- 1e4
-  hashes <- character(chunk_size + 1)
-  i <- 0
-  rawdata <- scan(
-    file = file, what = character(), nmax = -1, sep = "\n", quote = "",
-    skip = i * chunk_size, nlines = chunk_size, na.strings = "",
-    flush = FALSE, fill = FALSE, strip.white = FALSE, quiet = TRUE,
-    blank.lines.skip = FALSE, comment.char = "", allowEscapes = FALSE,
-    encoding = "UTF-8", skipNul = FALSE
-  )
-  while (length(rawdata)) {
-    hashes[1 + i %% chunk_size] <- hash(paste(hash(rawdata), collapse = "\n"))
-    i <- i + 1
-    if (i  %% chunk_size == 0) {
-      hashes[chunk_size + 1] <- hash(paste(hashes, collapse = "")) # nocov
-    }
+  if (file_test("-f", file)) {
+    chunk_size <- 1e4
+    hashes <- character(chunk_size + 1)
+    i <- 0
     rawdata <- scan(
       file = file, what = character(), nmax = -1, sep = "\n", quote = "",
       skip = i * chunk_size, nlines = chunk_size, na.strings = "",
       flush = FALSE, fill = FALSE, strip.white = FALSE, quiet = TRUE,
       blank.lines.skip = FALSE, comment.char = "", allowEscapes = FALSE,
       encoding = "UTF-8", skipNul = FALSE
+    )
+    while (length(rawdata)) {
+      hashes[1 + i %% chunk_size] <- hash(paste(hash(rawdata), collapse = "\n"))
+      i <- i + 1
+      if (i  %% chunk_size == 0) {
+        hashes[chunk_size + 1] <- hash(paste(hashes, collapse = "")) # nocov
+      }
+      rawdata <- scan(
+        file = file, what = character(), nmax = -1, sep = "\n", quote = "",
+        skip = i * chunk_size, nlines = chunk_size, na.strings = "",
+        flush = FALSE, fill = FALSE, strip.white = FALSE, quiet = TRUE,
+        blank.lines.skip = FALSE, comment.char = "", allowEscapes = FALSE,
+        encoding = "UTF-8", skipNul = FALSE
+      )
+    }
+  } else {
+    hashes <- sapply(
+      list.files(
+        file, pattern = "(index|[[:xdigit:]]{20}\\.tsv$)", full.names = TRUE
+      ),
+      datahash
     )
   }
   hash(paste(hashes, collapse = ""))