+++ tests/testthat/setup_test_data.R
@@ -17,7 +17,8 @@ test_data <- data.frame(
   test_logical = sample(c(TRUE, FALSE), size = test_n, replace = TRUE),
   test_POSIXct = as.POSIXct(
     sample(.Machine$integer.max, size = test_n, replace = TRUE),
-    origin = "1970-01-01"
+    origin = "1970-01-01",
+    tz = "UTC"
   ),
   test_Date = as.Date(
     c(sample(1e5, size = test_n - 1, replace = TRUE), 16000),
@@ -31,13 +32,11 @@ sorted_test_data <- test_data[order(test_data$test_Date), ]
 git2rdata:::set_local_locale(old_locale)
 sorted_test_data$test_character <- enc2utf8(sorted_test_data$test_character)
 rownames(sorted_test_data) <- NULL
-attr(sorted_test_data$test_POSIXct, "tzone") <- "UTC"
 
 test_subset <- head(test_data, ceiling(test_n / 2))
 
 sorted_test_subset <- test_subset[order(test_subset$test_Date), ]
 rownames(sorted_test_subset) <- NULL
-attr(sorted_test_subset$test_POSIXct, "tzone") <- "UTC"
 
 test_na <- test_data
 for (i in seq_along(test_na)) {
@@ -49,4 +48,3 @@ sorted_test_na <- test_na[
 ]
 git2rdata:::set_local_locale(old_locale)
 rownames(sorted_test_na) <- NULL
-attr(sorted_test_na$test_POSIXct, "tzone") <- "UTC"