+++ NEWS.md
@@ -1,3 +1,17 @@
+# git2rdata 0.3.0
+
+## New features
+
+* `write_vc()` gains an optional `split_by` argument.
+  See `vignette("split_by")` for more details.
+* `rename_variable()` efficiently renames variables in a stored `git2rdata`
+  object.
+
+## Bugfixes
+
+* `read_vc()`, `is_git2rdata()` and `is_git2rmeta()` now yield a better message
+  when both the data and metadata are missing.
+
 # git2rdata 0.2.2
 
 * Use the [checklist](https://inbo.github.io/checklist) package for CI.